Backend/requirements.txt
Django==6.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-django==4.7.0
django-filter==25.2
